test(desktop): keep artifact preview regression in adapter suite
